Made the children property of default props type to be optional
Setup the Global styles by including defaults and setting up the theme
classes for use in theme toggling later

Setup a Container component for grouping classes HTML elements and react
components and restricting the width of said objects

Added a utility file that exports a function that outputs the theme class based
on the passed boolean. If TRUE, outputs the light mode theme class,
otherwise outputs the dark mode theme class

Added useful types for use in theming and props in general. Width, and
Pixels type used in genering the proper Container component width
restriction, and general props type for use in components