Summary: Removed the previous implementation of, using theme context to pass the theme state to the toggler and using label and checkbox to determine the state, since it leads to a buggy theme switcher that changes themes on viewport resize.
Current implementation uses simple css classes and lifted state so the
theme is more localized. Additionally, it's more logical to let
react/next handle the state and just toggle classes instead of using a
lower level interface like label and checkboxes.

------------------------------------------------------------------------

Implemented a custom hook that will save the theme to localStorage
on theme change aside from generating a useState for the theme using the
previously saved local theme as initial value

------------------------------------------------------------------------

Removed the ThemeContext references and lifted the theme class to the
caller instead of by default

------------------------------------------------------------------------

Removed previous ThemeContext types and added types required for the
current theme implementation

------------------------------------------------------------------------

Renamed the previous theme class to theme-light

------------------------------------------------------------------------

Removed previous ThemeContext references and used the useTheme hook to
pass the state to the main container and the ThemeToggler

------------------------------------------------------------------------

Received the state from _app component and passed it to the theme
Toggler

------------------------------------------------------------------------

Removed the ThemeProvider component since context is not really needed
as the theme state is only passed 2 layer deep

------------------------------------------------------------------------

Implemented the ThemeToggler using button and span instead of using
labels, checkboxes and span